Added a new template for outputting a simple message: (simple_output.html). Currently index redirects there.
Added ability for users to log in after appropriate registration. They currently land on index.

Added very simplistic validation for username, password, and email for user registration.

Big! Added function "item_list", which uses render and template "list_template.html" to display a gridded list of items. Each item displayed must be of class ItemListObject. Created a function to take a list of ItemListObject(s) and create a 2d list so that there will be an appropriate number of rows passed to the template, and that each row will contain an appropriate number of ItemListObject(s).

On the to-do list:
(1) For every model, write a method allowing the data to be transformed into an ItemListObject.
(2) Write a view for showing users.
(3) Write a search to narrow now the users shown.
(4) Add to user profiles whether to be shown in search or not.

Lots to do!